+++ lib/Printer.ml
@@ -9,6 +9,17 @@ let printError src msg d =
     (Format.fprintf Format.std_formatter "@,@[%a@]" (Location.default_error_reporter ~src:(Some src)))
     []
 
+let printWarning src msg d =
+  Misc.Color.setup !Clflags.color ;
+  Code_frame.setup !Clflags.color ;
+  Format.fprintf Format.std_formatter "@[<v>@,  %a@,  %s@,@]"
+    (Location.print ~src:(Some src) ~message_kind:`warning "Lint Warning!")
+    Location.{loc_start= d.loc_start; loc_end= d.loc_end; loc_ghost= false}
+    msg ;
+  List.iter
+    (Format.fprintf Format.std_formatter "@,@[%a@]" (Location.default_error_reporter ~src:(Some src)))
+    []
+
 (* Location.default_error_reporter ?src:(Some (Some src)) Format.std_formatter
     Location.
       {loc= ; msg; sub= []; if_highlight= ""} *)