+++ lib/Rule.ml
@@ -13,6 +13,14 @@ type linter =
   | LintModuleBinding of (Parsetree.module_binding -> lintResult)
   | LintClassTypeDeclaration of (Parsetree.class_type_declaration -> lintResult)
 
+(* These are additional options that are parsed along side the rule name and the rule-specific options
+ * These rules are common to all rules and can be used to modify the behavior of the rule
+ *)
+module type LinterOptions = sig
+  (* Whether to show warnings for this rule - when false, or not specified, the rule will report as an error *)
+  val warning : bool
+end
+
 type meta = {ruleIdentifier: string; ruleName: string; ruleDescription: string}
 
 let meta_to_string meta =
@@ -20,6 +28,8 @@ let meta_to_string meta =
     meta.ruleDescription
 
 module type HASRULE = sig
+  include LinterOptions
+
   val meta : meta
 
   val linters : linter list